src/main/java/com/assignment/spring/config/OpenWeatherApiProperties.java
package com.assignment.spring.config;

import org.springframework.boot.context.properties.ConfigurationProperties;

@ConfigurationProperties(prefix = "app.openweather")
public class OpenWeatherApiProperties {
    private String url;
    private String appId;

    public String getUrl() {
        return url;
    }

    public void setUrl(String url) {
        this.url = url;
    }

    public String getAppId() {
        return appId;
    }

    public void setAppId(String appId) {
        this.appId = appId;
    }
}
